Conformity changes, part 2
