+++ PrimeAssemblyScript/solution_1/README.md
@@ -1,4 +1,6 @@
-# AssemblyScript implementation
+# AssemblyScript implementation by donmahallem
+
+![Category](https://img.shields.io/badge/Category-faithful-green)
 
 ## Run instructions
