+++ PrimeGo/solution_1/README.md
@@ -1,9 +1,13 @@
-# WHAT IS IT?
+# Go solution by bundgaard
+
+![Category](https://img.shields.io/badge/Category-faithful-green)
+
+## WHAT IS IT?
 Primes based on the code from Dave Plummer Primes repository.
 
 A Golang copy of the algorithm presented in PrimeCPP.
 
-# PERFORMANCE
+## PERFORMANCE
 
 Intel Xeon E3-1535M v6 3.1 GHz
 
@@ -13,7 +17,7 @@ Passes: 5796, Time: 5000590800, Avg: 0.000000, Limit: 1000000, Count1: 78498, Co
 
 
 
-# HOW TO BUILD
+## HOW TO BUILD
 
 Install Go 1.16.2
