+++ PrimeJava/solution_1/.dockerignore
@@ -1,0 +1,4 @@
+Dockerfile
+README.md
+gradlew.bat
+run-sieve.sh
