+++ PrimeJava/solution_1/.gitattributes
@@ -1,0 +1,3 @@
+* binary
+README.md text
+Dockerfile text
