+++ PrimeJava/solution_1/.gitignore
@@ -1,0 +1,3 @@
+build/
+.gradle/
+
