+++ PrimeJava/solution_1/Dockerfile
@@ -1,0 +1,8 @@
+FROM adoptopenjdk:11-jdk-hotspot
+
+WORKDIR /opt/app
+
+COPY src/main/java/PrimeSieveJava.java .
+RUN javac PrimeSieveJava.java
+
+ENTRYPOINT [ "java", "PrimeSieveJava" ]
