+++ PrimeJava/solution_1/README.md
@@ -1,0 +1,3 @@
+# Java solution by MansenC
+
+![Category](https://img.shields.io/badge/Category-faithful-green)
