+++ PrimeJava/solution_1/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+    id 'java-library'
+}
+
+jar {
+    manifest {
+        attributes(
+            'Class-Path': configurations.compile.collect { it.getName() }.join(' '),
+            'Main-Class': 'PrimeSieveJava'
+        )
+    }
+}
