+++ PrimeJava/solution_1/settings.gradle
@@ -1,0 +1,10 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * The settings file is used to specify which projects to include in your build.
+ * 
+ * Detailed information about configuring a multi-project build in Gradle can be found
+ * in the user guide at https://docs.gradle.org/5.0/userguide/multi_project_builds.html
+ */
+
+rootProject.name = 'PrimeSieveJava'
