--- PrimeJava/stl-faithful/MansenC/.gitignore
@@ -1,3 +1,0 @@
-build/
-.gradle/
-
