--- PrimeJava/stl-faithful/MansenC/build.gradle
@@ -1,12 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-jar {
-    manifest {
-        attributes(
-            'Class-Path': configurations.compile.collect { it.getName() }.join(' '),
-            'Main-Class': 'PrimeSieveJava'
-        )
-    }
-}
