+++ PrimeJulia/solution_1/README.md
@@ -1,0 +1,3 @@
+# Julia solution by dcbi
+
+![Category](https://img.shields.io/badge/Category-faithful-green)
