+++ PrimeLisp/solution_1/README.md
@@ -1,0 +1,3 @@
+# Lisp solution by mikehw
+
+![Category](https://img.shields.io/badge/Category-unfaithful-yellowgreen)
