+++ PrimeMIXAL/solution_1/.gitattributes
@@ -1,3 +1,2 @@
 *.sh binary
 *.mixal binary
-Dockerfile binary
