+++ PrimePHP/solution_1/Dockerfile
@@ -1,0 +1,7 @@
+FROM php:8.0-cli-alpine3.13
+
+WORKDIR /opt/app
+
+COPY *.php .
+
+ENTRYPOINT [ "php", "PrimePHP.php" ]
