+++ PrimePHP/solution_1/README.md
@@ -1,0 +1,3 @@
+# PHP solution by DennisdeBest
+
+![Category](https://img.shields.io/badge/Category-faithful-green)
