+++ PrimePython/stl-faithful/davepl/README.md
@@ -1,0 +1,3 @@
+# Python solution by davepl
+
+![Category](https://img.shields.io/badge/Category-faithful-green)
