+++ PrimeRuby/solution_1/.gitattributes
@@ -1,0 +1,1 @@
+*.rb binary
