+++ PrimeRuby/solution_1/README.md
@@ -1,6 +1,6 @@
 # Ruby solution by rbergen
 
-*Category: Faithful*
+![Category](https://img.shields.io/badge/Category-faithful-green)
 
 This is an implementation in Ruby.
